Fix issue about faulty pdf files
This commit will fix an issue where faulty pdf files break
the script. If fitz isn't able to open a pdf because its damaged,
it will write an error message and not return anything.
To catch that case, this commit implements an if clause in main
to check if create_pdf_info is returning something for the respective
pdf. If not an error message will be printed out into the console.